+++ Tests/Publish.CleanArchDapr.TestApplication/Publish.CleanArchDapr.TestApplication.Api/Controllers/DaprEventHandlerController.cs
@@ -1,12 +1,14 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
+using System.Transactions;
 using Dapr;
 using Intent.RoslynWeaver.Attributes;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.DependencyInjection;
 using Publish.CleanArchDapr.TestApplication.Application.Common.Eventing;
+using Publish.CleanArchDapr.TestApplication.Domain.Common.Interfaces;
 using Publish.CleanArchDapr.TestApplication.Eventing.Messages;
 
 [assembly: DefaultIntentManaged(Mode.Fully)]
@@ -21,11 +23,21 @@ namespace Publish.CleanArchDapr.TestApplication.Api.Controllers
     {
         private readonly ISender _mediatr;
         private readonly IServiceProvider _serviceProvider;
+        private readonly IEventBus _eventBus;
+        private readonly IDaprStateStoreUnitOfWork _daprStateStoreUnitOfWork;
+        private readonly IUnitOfWork _unitOfWork;
 
-        public DaprEventHandlerController(ISender mediatr, IServiceProvider serviceProvider)
+        public DaprEventHandlerController(ISender mediatr,
+            IServiceProvider serviceProvider,
+            IEventBus eventBus,
+            IDaprStateStoreUnitOfWork daprStateStoreUnitOfWork,
+            IUnitOfWork unitOfWork)
         {
             _mediatr = mediatr;
             _serviceProvider = serviceProvider;
+            _eventBus = eventBus ?? throw new ArgumentNullException(nameof(eventBus));
+            _daprStateStoreUnitOfWork = daprStateStoreUnitOfWork ?? throw new ArgumentNullException(nameof(daprStateStoreUnitOfWork));
+            _unitOfWork = unitOfWork ?? throw new ArgumentNullException(nameof(unitOfWork));
         }
 
         [HttpPost]
@@ -33,7 +45,17 @@ namespace Publish.CleanArchDapr.TestApplication.Api.Controllers
         public async Task HandleFullNamespaceEvent(FullNamespaceEvent @event, CancellationToken cancellationToken)
         {
             var handler = _serviceProvider.GetRequiredService<IIntegrationEventHandler<FullNamespaceEvent>>();
-            await handler.HandleAsync(@event, cancellationToken);
+
+            using (var transaction = new TransactionScope(TransactionScopeOption.Required,
+                new TransactionOptions { IsolationLevel = IsolationLevel.ReadCommitted }, TransactionScopeAsyncFlowOption.Enabled))
+            {
+                await handler.HandleAsync(@event, cancellationToken);
+                await _unitOfWork.SaveChangesAsync(cancellationToken);
+                transaction.Complete();
+            }
+
+            await _daprStateStoreUnitOfWork.SaveChangesAsync(cancellationToken);
+            await _eventBus.FlushAllAsync(cancellationToken);
         }
     }
 }